pyproject.toml
[tool.poetry]
name = "asli-car"
version = "0.1.0"
description = "Telegram bot used to scrape and notify about second hand cars from OLX which are listed by actual users and not dealers, which helps one to one dealing between the seller and buyer."
authors = ["Shivam Gupta <shivam.gupta02210@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
apscheduler = "^3.10.4"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.4"
requests = "^2.32.3"
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
